Added a copyConstructor method into Patient class.
Added a submit button at the end of  GetAllergy.razor to add the current Allergy variable into the List<Allergy> within Patient under a DataService.CurrentPatient.

Created a method for Dataservice that is intend to be used at the end of adding all of the Current patient's info. Will add CurrentPatient into a Patients List, erase value in CurrentPatient variable, and update/override the local xml file.