*Big Changes*
Created a Copy Constructor for GeneralInfo class.

Made SaveGeneralInfo() method in GetGeneralInfo, it prevents duplicate patients from being added into the Patients list by comparing, GeneralInfo.Name.FirstName and GeneralInfo.Name.LastName, to the existing patients (ignoring the case). If it passes it gets allowed to be saved.

Added a statement, in AtEndSubmitAndAddPatient(), which checks whether or not a local directory exists, else it creates one.

In DataService made the method LoadExistingPatients(), it checks to see if the patient list file exists in the local directory. If true, it'll assign the stored values into a Patient list variable.

In the StartUpPage added a load button binded with LoadExistingPatients().

in StartPage, changed the foreach loop, that shows the patient list, into a for loop making it easier to index and escalate. Added a Delete button, to remove a patient indicated by an 'X', from the List of Patients variable.