In Patient Class, a string PatientID.
GeneralInfo class, changed DateTime DateOfBirth to DateOnly.

Made a method to create a string ID for the patient. It concatenates the first two letters of the first name and last name (lowercase) and adds birth date to the end of the string.
It gets called and assigned onto the PatientID string of the CurrentPatient, that is being created. Right before it gets added onto the List<Patient> and saved.