In ListOfSavedPatients, replaced the For loop with a Foreach loop (the indicator for the DeletePatient method wasn't needed anymore). Using Patient.ID as a delete reference.
Created a GetCurrentPatient method to make a deep copy of the CurrentPatient and return it.

Changed the DeletePatient method to check if the parameter's input exists as a patient.ID, then it removes All patients with that value (just in case of duplicates), returns true if successful, saves the new list of patients, or it returns false if no patients were deleted.