In DataService, created a method, intended for the initial setup, for an autocomplete feature.
The goal of this method is to ensure the availability of the wordlists that are to be used in an autocomplete feature. The method AutoCompleteWordListInitialSetup is made for the intial setup. It checks if both the filepaths and lists exist locally, are empty, or null, and if not then it overrides them using specific methods.